+++ src/Screens/CityManagerPanels/CityProduction.cs
@@ -145,7 +145,7 @@ namespace CivOne.Screens.CityManagerPanels
 		private bool Buy()
 		{
 			string name = (_city.CurrentProduction as ICivilopedia)!.Name;
-			short playerGold = Game.CurrentPlayer.Gold;
+			int playerGold = Game.CurrentPlayer.Gold;
 			short buyPrice = _city.BuyPrice;
 			if (buyPrice <= 0)
 				return true;
